firefox now has optional site permissions
