+++ scripts/build-src.ts
@@ -57,13 +57,10 @@ function src(
     );
   const bundleOptionsScript = (): Promise<esbuild.BuildResult> =>
     bundleJS(distPath, path.resolve(srcPath, 'ui', 'options', 'options.ts'));
-  const bundleBackgroundScript = (): Promise<esbuild.BuildResult> =>
-    bundleJS(distPath, path.resolve(srcPath, 'background', 'background.ts'));
 
   const bundleAll: Promise<esbuild.BuildResult> = bundleMainScript()
     .then(bundlePopupScript)
-    .then(bundleOptionsScript)
-    .then(bundleBackgroundScript);
+    .then(bundleOptionsScript);
 
   const copyPopup: Promise<void[]> = Promise.all(
     [
