+++ src/ui/popup/settings-popup.css
@@ -4,6 +4,8 @@ body {
   color: #1a202c;
   font-size: 16px;
   font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+  /** firefox has bigger popup width limit, this is to bring it to parity with chromium **/
+  max-width: 300px;
 }
 
 #content {
